fix: Curator CLI のセッション状態に failed_themes と category_distribution を追加
CLI (`python -m curator_agents`) が `{existing_titles}` しかセッション状態にセットしておらず、
`{category_distribution}` と `{failed_themes}` がリテラル文字列のまま LLM に渡っていた問題を修正。

- `curator_agents/queries.py` を新規作成し、ドメインクエリ関数を一元化
- `services/curator.py` から重複コードを削除して `queries.py` から import
- `cli.py` に `failed_themes` と `category_distribution` のセッション状態セットを追加
- テストのパッチ先を `curator_agents.queries.*` に更新
