+++ services/curator.py
@@ -11,7 +11,6 @@ Endpoints:
 
 import json
 import os
-from collections import Counter
 
 from fastapi import FastAPI
 from fastapi.responses import JSONResponse
@@ -22,85 +21,15 @@ from google.adk.sessions import InMemorySessionService
 from google.genai import types
 
 from curator_agents.agents.curator import curator_agent
+from curator_agents.queries import (
+    get_existing_titles,
+    get_category_distribution,
+    format_category_distribution,
+)
 from translator_agents.agents.translator import translator_agent
-from shared.firestore import get_firestore_client
 from shared.pipeline_failure import get_recent_failures
 
 app = FastAPI()
-
-# 8分類コードの定義（mystery_id プレフィックス）
-ALL_CATEGORIES = ["HIS", "FLK", "ANT", "OCC", "URB", "CRM", "REL", "LOC"]
-
-
-def get_existing_titles() -> list[str]:
-    """Fetch titles of all existing mysteries from Firestore."""
-    try:
-        db = get_firestore_client()
-        docs = db.collection("mysteries").select(["title"]).stream(timeout=10)
-        return [
-            doc.to_dict().get("title", "")
-            for doc in docs
-            if doc.to_dict().get("title")
-        ]
-    except Exception as e:
-        print(f"Warning: Could not fetch existing titles: {e}")
-        return []
-
-
-def get_category_distribution() -> dict[str, int]:
-    """Firestore の mystery_id フィールドからカテゴリ分布を集計する。
-
-    mystery_id は {CLS}-{ST}-{AREA}-{TS} 形式で、先頭3文字が分類コード。
-    """
-    try:
-        db = get_firestore_client()
-        docs = db.collection("mysteries").select(["mystery_id"]).stream(timeout=10)
-        prefixes = []
-        for doc in docs:
-            mystery_id = doc.to_dict().get("mystery_id", "")
-            if mystery_id and len(mystery_id) >= 3:
-                prefix = mystery_id[:3].upper()
-                if prefix in ALL_CATEGORIES:
-                    prefixes.append(prefix)
-        return dict(Counter(prefixes))
-    except Exception as e:
-        print(f"Warning: Could not fetch category distribution: {e}")
-        return {}
-
-
-def format_category_distribution(distribution: dict[str, int]) -> str:
-    """カテゴリ分布をプロンプト用テキストに変換する。
-
-    空の場合はコールドスタート向けメッセージを返す。
-    データありの場合は各カテゴリの件数と過小表現カテゴリを表示。
-    """
-    if not distribution:
-        return (
-            "No existing articles yet. This is a fresh start — "
-            "use all 8 categories broadly. Aim for maximum diversity across "
-            "HIS, FLK, ANT, OCC, URB, CRM, REL, and LOC."
-        )
-
-    total = sum(distribution.values())
-    lines = []
-    for cat in ALL_CATEGORIES:
-        count = distribution.get(cat, 0)
-        lines.append(f"  {cat}: {count} article(s)")
-
-    # 過小表現カテゴリ（0件 or 平均以下）の特定
-    avg = total / len(ALL_CATEGORIES)
-    underrepresented = [
-        cat for cat in ALL_CATEGORIES
-        if distribution.get(cat, 0) < avg
-    ]
-
-    if underrepresented:
-        lines.append(
-            f"\nUnderrepresented categories (prioritize these): "
-            f"{', '.join(underrepresented)}"
-        )
-
-    return "\n".join(lines)
 
 
 async def run_curator() -> dict:
